devtools/client/shared/components/test/mochitest/test_reps_function.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test Func rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - Func</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  let { Rep } = browserRequire("devtools/client/shared/components/reps/rep");
  let { Func } = browserRequire("devtools/client/shared/components/reps/function");
  const { MODE } = browserRequire("devtools/client/shared/components/reps/constants");

  const componentUnderTest = Func;

  try {
    // Test that correct rep is chosen
    const gripStub = getGripStub("testNamed");
    const renderedRep = shallowRenderComponent(Rep, { object: gripStub });
    is(renderedRep.type, Func.rep, `Rep correctly selects ${Func.rep.displayName}`);

    yield testNamed();
    yield testVarNamed();
    yield testAnon();
    yield testLongName();
    yield testAsyncFunction();
    yield testAnonAsyncFunction();
    yield testGeneratorFunction();
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function testNamed() {
    // Test declaration: `function testName() { let innerVar = "foo" }`
    const testName = "testNamed";

    const defaultOutput = "function testName()";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testUserNamed() {
    // Test declaration: `function testName() { let innerVar = "foo" }`
    const testName = "testUserNamed";

    const defaultOutput = "function testUserName()";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testVarNamed() {
    // Test declaration: `let testVarName = function() { }`
    const testName = "testVarNamed";

    const defaultOutput = "function testVarName()";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testAnon() {
    // Test declaration: `() => {}`
    const testName = "testAnon";

    const defaultOutput = "function ()";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testLongName() {
    // Test declaration: `let f = function loooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong() { }`
    const testName = "testLongName";

    const defaultOutput = "function looooooooooooooooooooooooooooooooooooooooooooooooo\u2026ooooooooooooooooooooooooooooooooooooooooooooong()";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      }
    ];

    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testAsyncFunction() {
    const testName = "testAsyncFunction";

    const defaultOutput = "async function waitUntil2017()";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];
    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testAnonAsyncFunction() {
    const testName = "testAnonAsyncFunction";

    const defaultOutput = "async function ()";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];
    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testGeneratorFunction() {
    const testName = "testGeneratorFunction";

    const defaultOutput = "function* fib()";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];
    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function testAnonGeneratorFunction() {
    const testName = "testAnonGeneratorFunction";

    const defaultOutput = "function* ()";

    const modeTests = [
      {
        mode: undefined,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.TINY,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.SHORT,
        expectedOutput: defaultOutput,
      },
      {
        mode: MODE.LONG,
        expectedOutput: defaultOutput,
      }
    ];
    testRepRenderModes(modeTests, testName, componentUnderTest, getGripStub(testName));
  }

  function getGripStub(functionName) {
    switch (functionName) {
      case "testNamed":
        return {
          "type": "object",
          "class": "Function",
          "actor": "server1.conn6.obj35",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "isAsync": false,
          "isGenerator": false,
          "name": "testName",
          "displayName": "testName",
          "location": {
            "url": "debugger eval code",
            "line": 1
          }
        };

      case "testUserNamed":
        return {
          "type": "object",
          "class": "Function",
          "actor": "server1.conn6.obj35",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "isAsync": false,
          "isGenerator": false,
          "name": "testName",
          "userDisplayName": "testUserName",
          "displayName": "testName",
          "location": {
            "url": "debugger eval code",
            "line": 1
          }
        };

      case "testVarNamed":
        return {
          "type": "object",
          "class": "Function",
          "actor": "server1.conn7.obj41",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "isAsync": false,
          "isGenerator": false,
          "displayName": "testVarName",
          "location": {
            "url": "debugger eval code",
            "line": 1
          }
        };

      case "testAnon":
        return {
          "type": "object",
          "class": "Function",
          "actor": "server1.conn7.obj45",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "isAsync": false,
          "isGenerator": false,
          "location": {
            "url": "debugger eval code",
            "line": 1
          }
        };

      case "testLongName":
        return {
          "type": "object",
          "class": "Function",
          "actor": "server1.conn7.obj67",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "isAsync": false,
          "isGenerator": false,
          "name": "loooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong",
          "displayName": "loooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong",
          "location": {
            "url": "debugger eval code",
            "line": 1
          }
        };


      case "testAsyncFunction":
        return {
          "type": "object",
          "class": "Function",
          "actor": "server1.conn7.obj45",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "isAsync": true,
          "isGenerator": false,
          "name": "waitUntil2017",
          "displayName": "waitUntil2017",
          "location": {
            "url": "debugger eval code",
            "line": 1
          }
        };

      case "testAnonAsyncFunction":
        return {
          "type": "object",
          "class": "Function",
          "actor": "server1.conn7.obj45",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "isAsync": true,
          "isGenerator": false,
          "location": {
            "url": "debugger eval code",
            "line": 1
          }
        };

      case "testGeneratorFunction":
        return {
          "type": "object",
          "class": "Function",
          "actor": "server1.conn7.obj45",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "isAsync": false,
          "isGenerator": true,
          "name": "fib",
          "displayName": "fib",
          "location": {
            "url": "debugger eval code",
            "line": 1
          }
        };

      case "testAnonGeneratorFunction":
        return {
          "type": "object",
          "class": "Function",
          "actor": "server1.conn7.obj45",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "isAsync": false,
          "isGenerator": true,
          "location": {
            "url": "debugger eval code",
            "line": 1
          }
        };
    }
  }
});
</script>
</pre>
</body>
</html>
